requirements.txt
asgiref==3.2.10
Django==3.0.8
Markdown==3.2.2
psycopg2==2.8.5
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
